pyproject.toml
[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[project]
name = "sota-model"
version = "0.1.0"
description = "SOTA dense foundation model targeting SuperModel 4.7-class capabilities (see )"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Ashish" }]

dependencies = [
    "torch>=2.4",
    "transformers>=4.45",
    "tokenizers>=0.20",
    "datasets>=2.20",
    "safetensors>=0.4.5",
    "numpy>=1.26",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
train = [
    "deepspeed>=0.15",
    "accelerate>=0.34",
    "wandb>=0.18",
    "ray[default]>=2.30",
]
flash = [
    "flash-attn>=2.6; platform_system == 'Linux'",
]
serve = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sse-starlette>=2.1",
    "prometheus-client>=0.20",
]
dev = [
    "pytest>=8.3",
    "pytest-xdist>=3.6",
    "ruff>=0.6",
    "mypy>=1.11",
]

[project.scripts]
sota-pretrain = "sota_model.training.pretrain:main"
sota-serve = "sota_model.serving.server:main"
sota-eval = "sota_model.evaluation.run:main"

[tool.hatch.build.targets.wheel]
packages = ["src/sota_model"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"
